Reduce memory footprint per histogram.

+++ tests/data_access.rs
@@ -28,7 +28,8 @@ struct Loaded {
 }
 
 const TRACKABLE_MAX: u64 = 3600 * 1000 * 1000;
-const SIGFIG: u32 = 3;
+// Store up to 2 * 10^3 in single-unit precision. Can be 5 at most.
+const SIGFIG: u8 = 3;
 const EINTERVAL: u64 = 10000; /* 10 msec expected EINTERVAL */
 const SCALEF: u64 = 512;
